relocate ui model
