--- common/src/main/kotlin/org/wycliffeassociates/otter/common/data/ConflictResolution.kt
@@ -1,6 +1,0 @@
-package org.wycliffeassociates.otter.common.data
-
-enum class ConflictResolution {
-    DISCARD,
-    OVERRIDE
-}
