+++ jvm/workbookapp/src/main/kotlin/org/wycliffeassociates/otter/jvm/workbookapp/ui/model/ConflictResolution.kt
@@ -1,0 +1,6 @@
+package org.wycliffeassociates.otter.jvm.workbookapp.ui.model
+
+enum class ConflictResolution {
+    DISCARD,
+    OVERRIDE
+}
